Add fast catchup goal test
