+++ test/e2e-go/cli/goal/expect/goalTxValidityTest.exp
@@ -46,6 +46,9 @@ proc TestLastValidInTx { CMD TX_FILE EXPECTED_LAST_VALID } {
 
 if { [catch {
     # Create network
+    ::AlgorandGoal::CreateNetwork $NETWORK_NAME $NETWORK_TEMPLATE $TEST_ALGO_DIR $TEST_ROOT_DIR
+
+    # Start network
     ::AlgorandGoal::StartNetwork $NETWORK_NAME $NETWORK_TEMPLATE $TEST_ALGO_DIR $TEST_ROOT_DIR
 
     # use goal node status command to wait for round 0
