+++ test/e2e-go/features/catchup/catchpointCatchup_test.go
@@ -14,17 +14,14 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with go-algorand.  If not, see <https://www.gnu.org/licenses/>.
 
-package rewards
+package catchup
 
 import (
-	"fmt"
-	"net"
 	"net/http"
 	"os"
 	"os/exec"
 	"path/filepath"
 	"runtime"
-	"strings"
 	"syscall"
 	"testing"
 	"time"
@@ -144,7 +141,7 @@ func TestBasicCatchpointCatchup(t *testing.T) {
 	primaryListeningAddress, err := primaryNode.GetListeningAddress()
 	a.NoError(err)
 
-	wp, err := makeWebProxy(primaryListeningAddress, func(response http.ResponseWriter, request *http.Request, next http.HandlerFunc) {
+	wp, err := fixtures.MakeWebProxy(primaryListeningAddress, func(response http.ResponseWriter, request *http.Request, next http.HandlerFunc) {
 		// prevent requests for block #2 to go through.
 		if request.URL.String() == "/v1/test-v1/block/2" {
 			response.WriteHeader(http.StatusBadRequest)
@@ -204,112 +201,4 @@ func TestBasicCatchpointCatchup(t *testing.T) {
 
 	secondNode.StopAlgod()
 	primaryNode.StopAlgod()
-}
-
-type webProxyInterceptFunc func(http.ResponseWriter, *http.Request, http.HandlerFunc)
-
-type webProxy struct {
-	server      *http.Server
-	listener    net.Listener
-	destination string
-	intercept   webProxyInterceptFunc
-}
-
-func makeWebProxy(destination string, intercept webProxyInterceptFunc) (wp *webProxy, err error) {
-	if strings.HasPrefix(destination, "http://") {
-		destination = destination[7:]
-	}
-	wp = &webProxy{
-		destination: destination,
-		intercept:   intercept,
-	}
-	wp.server = &http.Server{
-		Handler: wp,
-	}
-	wp.listener, err = net.Listen("tcp", "localhost:")
-	if err != nil {
-		return nil, err
-	}
-	go func() {
-		wp.server.Serve(wp.listener)
-	}()
-	return wp, nil
-}
-
-func (wp *webProxy) GetListenAddress() string {
-	return wp.listener.Addr().String()
-}
-
-func (wp *webProxy) Close() {
-	// we can't use shutdown, since we have tunneled websocket, which is a hijacked connection
-	// that http.Server doens't know how to handle.
-	wp.server.Close()
-}
-
-func (wp *webProxy) ServeHTTP(response http.ResponseWriter, request *http.Request) {
-	//fmt.Printf("incoming request for %v\n", request.URL)
-	if wp.intercept == nil {
-		wp.Passthrough(response, request)
-		return
-	}
-	wp.intercept(response, request, wp.Passthrough)
-}
-
-func (wp *webProxy) Passthrough(response http.ResponseWriter, request *http.Request) {
-	client := http.Client{}
-	clientRequestURL := *request.URL
-	clientRequestURL.Scheme = "http"
-	clientRequestURL.Host = wp.destination
-	clientRequest, err := http.NewRequest(request.Method, clientRequestURL.String(), request.Body)
-	if err != nil {
-		fmt.Printf("Passthrough request assembly error %v (%#v)\n", err, clientRequestURL)
-		response.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	if request.Header != nil {
-		for headerKey, headerValues := range request.Header {
-			for _, headerValue := range headerValues {
-				clientRequest.Header.Add(headerKey, headerValue)
-			}
-		}
-	}
-	clientResponse, err := client.Do(clientRequest)
-	if err != nil {
-		fmt.Printf("Passthrough request error %v (%v)\n", err, request.URL.String())
-		response.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	if clientResponse.Header != nil {
-		for headerKey, headerValues := range clientResponse.Header {
-			for _, headerValue := range headerValues {
-				response.Header().Add(headerKey, headerValue)
-			}
-		}
-	}
-	response.WriteHeader(clientResponse.StatusCode)
-	ch := make(chan []byte, 10)
-	go func(outCh chan []byte) {
-		defer close(outCh)
-		if clientResponse.Body == nil {
-			return
-		}
-		defer clientResponse.Body.Close()
-		for {
-			buf := make([]byte, 4096)
-			n, err := clientResponse.Body.Read(buf)
-			if n > 0 {
-				outCh <- buf[:n]
-			}
-			if err != nil {
-				break
-			}
-
-		}
-	}(ch)
-	for bytes := range ch {
-		response.Write(bytes)
-		if flusher, has := response.(http.Flusher); has {
-			flusher.Flush()
-		}
-	}
 }
